+++ lib/domain/mixin/connection_restore_mixin.dart
@@ -4,7 +4,7 @@ import 'package:application_base/core/service/service_locator.dart';
 import 'package:application_base/domain/subject/network_subject.dart';
 import 'package:meta/meta.dart';
 
-// TODO(Alex): перенести в domain layer
+///
 base mixin ConnectionRestoreMixin {
   ///
   late StreamSubscription<NetworkEvent> _subscriptionConnection;